fix: clear the compiler warnings on Linux and Windows
gcc had two, MSVC fourteen, and none of them were noise:

- The HTTP connection pool keys itself on "host:port" in a buffer as wide
  as the host field alone, so gcc warned that the port could be truncated
  away. Two hosts differing only past that point would then share one pool.
  The host width and the key width are now named and the key is the wider
  of the two.
- CMUTIL_SocketConnectIPCInternal passed a struct sockaddr_in* where a
  CMUTIL_SocketAddr* was expected. Same address either way, so the code
  worked, but the declared types disagreed - it now passes the storage it
  meant to.
- select() takes an int nfds, and a SOCKET is a 64 bit handle on Windows
  where that argument is ignored outright; connect() and bind() take a
  socklen_t, not a size_t; Socket::Write and ByteBuffer::AddBytes take a
  uint32_t, not a size_t. All now say so.
- crypto_test declared an EVP_CIPHER it never used.

+++ test/dgram_test.c
@@ -92,7 +92,7 @@ int main() {
     CMLogInfo("datagram socket connected to %s:%d", rhost, rport);
 
     const char *send_data = "this is sample message";
-    CMCall(buffer, AddBytes, (const uint8_t*)send_data, strlen(send_data));
+    CMCall(buffer, AddBytes, (const uint8_t*)send_data, (uint32_t)strlen(send_data));
     sr = CMCall(client, Send, buffer, 1000);
     ASSERT(sr == CMSocketOk, "Send");
 
@@ -108,7 +108,7 @@ int main() {
 
     client = CMUTIL_DGramSocketCreateBind(NULL);
     send_data = "second data";
-    CMCall(buffer, AddBytes, (const uint8_t*)send_data, strlen(send_data));
+    CMCall(buffer, AddBytes, (const uint8_t*)send_data, (uint32_t)strlen(send_data));
     sr = CMCall(client, SendTo, buffer, &localaddr, 1000);
     ASSERT(sr == CMSocketOk, "SendTo");
 
@@ -127,7 +127,7 @@ int main() {
 
     CMCall(buffer, Clear);
     send_data = "!@#$$#@!";
-    CMCall(buffer, AddBytes, (const uint8_t*)send_data, strlen(send_data));
+    CMCall(buffer, AddBytes, (const uint8_t*)send_data, (uint32_t)strlen(send_data));
     sr = CMCall(client, SendTo, buffer, &localaddr, 1000);
     if (sr != CMSocketOk) {
         CMLogError("datagram socket send failed - failed");
